Finished debugging final AST draft
AST should now create a proper tree. Also includes complete example.

+++ P4 Project/P4 Project/AST/Expressions/Values/NumConst.cs
@@ -8,11 +8,12 @@ namespace P4_Project.AST.Expressions.Values
 {
     class NumConst : ExprNode
     {
-        int val;
+        double val;
 
-        public NumConst(int _val)
+        public NumConst(double _val)
         {
             val = _val;
         }
+
     }
 }
